feat: agora existem avisos sempre que nao houverem itens disponiveis

+++ src/view/StockManagement.java
@@ -1,9 +1,7 @@
 package view;
 
-import db.Db;
 import model.dao.DaoFactory;
 import model.dao.StockDao;
-import model.dao.impl.StockDaoJDBC;
 import model.entities.Stock;
 
 import java.util.List;
@@ -15,8 +13,12 @@ public class StockManagement {
         StockDao stockDaoJDBC = new DaoFactory().createStockDao();
 
         List<Stock> stocks = stockDaoJDBC.findAll();
+        if(stocks.size() > 0){
+            stocks.forEach(stock -> System.out.println( stock.toString()));
+        }else{
+            System.out.println("Unfortunally we don't have any items in stock");
+        }
 
-        stocks.forEach(stock -> System.out.println( stock.toString()));
     }
 
     public static void addStock(){
@@ -45,12 +47,19 @@ public class StockManagement {
 
         int id ;
         List<Stock> stocks = stockDaoJDBC.findAll();
-        stocks.forEach(stock -> System.out.println( stock.toString()));
-        System.out.println("Choose the product´s id to be deleted: ");
-        id = sc.nextInt();
+        if(stocks.size() > 0){
 
-        stockDaoJDBC.deleteById(id);
-        System.out.println("Item deleted");
+            stocks.forEach(stock -> System.out.println( stock.toString()));
+            System.out.println("Choose the product´s id to be deleted: ");
+            id = sc.nextInt();
+
+            stockDaoJDBC.deleteById(id);
+            System.out.println("Item deleted");
+
+
+        }else{
+            System.out.println("Unfortunally we don't have any items in stock");
+        }
 
 
 
@@ -60,45 +69,52 @@ public class StockManagement {
         Scanner sc = new Scanner(System.in);
         StockDao stockDaoJDBC = new DaoFactory().createStockDao();
         List<Stock> stocks = stockDaoJDBC.findAll();
-        stocks.forEach(stock -> System.out.println( stock.toString()));
 
-        Stock aux = new Stock();
-        int id;
+        if(stocks.size() > 0){
+            stocks.forEach(stock -> System.out.println( stock.toString()));
+            Stock aux = new Stock();
+            int id;
 
-        System.out.println("Choose the product´s id to be updated: ");
-        id = sc.nextInt();
+            System.out.println("Choose the product´s id to be updated: ");
+            id = sc.nextInt();
 
-        for (Stock stock : stocks) {
-            if (stock.getId() == id) {
-                aux = stock;
+            for (Stock stock : stocks) {
+                if (stock.getId() == id) {
+                    aux = stock;
+                }
             }
-        }
 
 
-        System.out.println("Write the product´s name (if you dont want to update, just leave it blank ): ");
-        sc.nextLine();
-        String name = sc.nextLine();
-        if (!name.trim().isBlank()){
-        aux.setName(name);
-        }
-        System.out.println("Write the product's stock quantity. (if you dont want to update, just leave it blank): ");
-        String quantity = sc.nextLine();
-        if(!quantity.trim().isBlank()){
-            aux.setQuantity(Integer.parseInt(quantity));
-        }
-        System.out.println(" Write the product's price. (if you dont want to update, just leave it blank): ");
-        String price = sc.nextLine();
-        if(!price.trim().isBlank()){
-            aux.setPrice(Double.parseDouble(price));
-        }
-        System.out.println(" Write the product's category. (if you dont want to update, just leave it blank): ");
-        String category = sc.nextLine();
-        if(!category.trim().isBlank()){
-            aux.setCategory(category);
+            System.out.println("Write the product´s name (if you dont want to update, just leave it blank ): ");
+            sc.nextLine();
+            String name = sc.nextLine();
+            if (!name.trim().isBlank()){
+                aux.setName(name);
+            }
+            System.out.println("Write the product's stock quantity. (if you dont want to update, just leave it blank): ");
+            String quantity = sc.nextLine();
+            if(!quantity.trim().isBlank()){
+                aux.setQuantity(Integer.parseInt(quantity));
+            }
+            System.out.println(" Write the product's price. (if you dont want to update, just leave it blank): ");
+            String price = sc.nextLine();
+            if(!price.trim().isBlank()){
+                aux.setPrice(Double.parseDouble(price));
+            }
+            System.out.println(" Write the product's category. (if you dont want to update, just leave it blank): ");
+            String category = sc.nextLine();
+            if(!category.trim().isBlank()){
+                aux.setCategory(category);
 
+            }
+
+            stockDaoJDBC.update(aux);
+
+
+        }else{
+            System.out.println("Unfortunally we don't have any items in stock");
         }
 
-        stockDaoJDBC.update(aux);
 
 
     }
